Add initial projects with basic UI flows
MobileDirectory is the main application and will be the only package
that users install upon completion of the project.  BetaManager will
perform various beta related tasks such as registering devices,
reporting issues, checking for updates, and running unit tests.

Upon startup, MobileDirectory will check if BetaManager is installed,
and if it is, it will enable all beta features in the application.

Signed-off-by: Kevin Wells <wellska1@rose-hulman.edu>