Render basic text overlays of buildings
This change does render text names of buildings, but it is missing some
key functionality.  It does not make calls to the actual webservice and
is instead using hardcoded data that is written to the database on
startup.  Also, it is not detecting text collisions when the map is
zoomed out.

The server data will need a bit for each building indicating if the
Google Map satellite hybrid view contains a visible name for a given
building so that the client can know to not display that building name.